+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of grafana.
 
 - Add `enabled` property to `config_unified_alerting` resource [grafana docs](https://grafana.com/docs/grafana/latest/alerting/migrating-alerts/roll-back/)
 - Add `force_migration` property to `config` resource to allow unified alerting to be rolled back
+- Migrate install resource to use new default repo URLs
 
 ## 10.1.0 - *2022-10-13*